requirements.txt
numpy==1.26.3
pandas==2.1.4
pillow==10.2.0
prompt-toolkit==3.0.43
python-dateutil==2.8.2
pytz==2023.3.post1
six==1.16.0
tzdata==2023.4
wcwidth==0.2.12
